✅ Moved hyrdation logic to header
